+++ src/test/kotlin/klib/extensions/TByte.kt
@@ -1,0 +1,14 @@
+package klib.extensions
+
+import com.atvgstudios.klib.extensions.mul
+import org.junit.Test
+
+fun b(num: Int): Byte = num.toByte()
+
+class TByte {
+    @Test
+    fun mul() {
+        val byteResult = b(0x02) mul b(0x02)
+        assert(byteResult == b(0x04))
+    }
+}